rest update added

+++ Billing-App/src/main/java/com/nayan/model/BillingModel.java
@@ -1,11 +1,6 @@
 package com.nayan.model;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.Table;
-import javax.xml.bind.annotation.XmlRootElement;
 
 import org.springframework.data.annotation.Transient;
 import org.springframework.data.mongodb.core.mapping.Document;
@@ -17,20 +12,20 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-@Document(collection="BILL")
+@Document(collection = "BILL")
 
 public class BillingModel {
 
 //	private int billId;
 //	@OneToMany(mappedBy = "bill")
 //	private List<BillDTO> list;
-	
-	 @Transient
-	 public static final String SEQUENCE_NAME = "users_sequence";
+
+	@Transient
+	public static final String SEQUENCE_NAME = "users_sequence";
 	@Id
-	
+
 	private long id;
 	private int billId;
-	private int price;
+	private float price;
 	private String name;
 }
